Fix another surface interval NaN case
In the case that a repeat dive is already bad, our algorithm will
return an impossible result, i.e. NaN. We have to account for this
case by setting the surface interval as high as possible. Although
the repeat dive will still be bad, this makes sense because the
user should adjust it.